MyProject/requirements.txt
pytest~=8.2.2
PyYAML~=6.0.1
pytest-cov